group coding on 10/25

+++ controllers/adminController.js
@@ -9,152 +9,155 @@ const jwt = require('jsonwebtoken')
 
 const { JWT_KEY_SECRET } = require("../config")
 
-
-const homePage = (req, res) => {
-    res.render("home")
-
-}
-
 // NEW - SEND FORM
 const sendNewAdminForm = (req, res, next) => {
-    let accessGranted = false
-    if(req.cookies.access_token) {
-        accessGranted = true
+    let entryAccess = false
+    if(req.cookies.admintoken) {
+        entryAccess = true
     } 
-    res.render('admin/newadmin.ejs', {accessGranted})
+    res.render('admin/newadmin.ejs', {entryAccess})
 }
 
-// NEW - POST NEW ADMIN
-const createNewAdmin = (req, res, next) => {
-    const requiredFields = ['email', 'password', 'fullname', 'organization', 'username'];
 
-    for(let i=0; i < requiredFields.length; i++) {
-        const field = requiredFields[i]
+
+// // NEW - POST NEW ADMIN
+const createNewAdmin = async (req, res) => {
+    const requiredFields = ['email', 'password', 'fullname', 'organization'];
+    for(let field of requiredFields) {
         if(!(field in req.body)) {
             const errorMessage = `missing ${field} in request body`;
             return res.send(errorMessage);
     }}
-
     req.body.email = req.body.email.toLowerCase()
-    const { email, password, fullname, organization, username } = req.body
-
-    bcrypt.hash(password, 12)
-        .then(encryptedPw => {
-            const newAdmin = { 
-                email, password: encryptedPw, fullname, organization, username }
-            Admin.create(newAdmin)
-                .then(admin => {
-                    const token = jwt.sign(
-                        { userId: admin.id, email: admin.email },
-                        JWT_KEY_SECRET)
-                        return res.cookie('access_token', token).redirect('/admin/login')  
-                })
-    })
+    const { email, password, fullname, organization } = req.body
+    try {
+        const encryptedPw = await bcrypt.hash(password, 12)
+        const newAdmin = { email, password: encryptedPw, fullname, organization }
+        const admin = Admin.create(newAdmin)
+        const token = jwt.sign(
+            { userId: admin.id, email: admin.email },
+            JWT_KEY_SECRET,
+            { expiresIn: '1hr' }
+            );
+        return res.cookie('admintoken', token).redirect('/admin/home')  
+    } catch (error) {
+        console.error(error)
+    }
 }
-
 
 
 // LOGIN - GET
 const sendLoginForm = (req, res, next) => {
-    let accessGranted = false
-    if(req.cookies.access_token) {
-        accessGranted = true
+    let entryAccess = false;
+    if(req.cookies.admintoken) {
+        entryAccess = true
     } 
-    res.render('admin/login.ejs', {accessGranted})
+    res.render('admin/login.ejs', {entryAccess})
 }
 
 // LOGIN - POST 
-const adminLogin = (req, res) => {
-    req.body.email.toLowerCase();
-    Admin.findOne({ email: req.body.email })
-        .then((admin) => {
-            if (!admin){
-                return res.send('email not found')
-            }
-            bcrypt.compare(req.body.password, admin.password)
-                .then((matched) => {
-                    if (matched === false) {
-                        return res.send('invalid password, try again')
-                    }
-                    const token = jwt.sign(
-                        { userId: admin.id, email: admin.email },
-                        JWT_KEY_SECRET)
-                    return res.cookie('accesstoken', token)
-                        .redirect("/admin/home")
-                })
-        })
+const adminLogin = async (req, res, next) => {
+    try{
+        req.body.email.toLowerCase();
+        const admin = await Admin.findOne({ email: req.body.email })
+        console.log({ email: req.body.email })
+        if (!admin){
+            return res.send('email not found')
+        }
+        const verify = bcrypt.compare(req.body.password, admin.password);
+        if (!verify) {
+            return res.send("invalid password")
+        }
+        const token = jwt.sign(
+            { userId: admin.id, email: admin.email },
+            JWT_KEY_SECRET,
+            { expiresIn: '1hr' }
+        );
+        return res.cookie('admintoken', token).redirect("/admin/home")
+    } catch (error) {
+        console.error(error)
+    }
 }
 
 // LOGOUT
 const adminLogout = (req, res, next) => {
-    const token = req.cookies.accesstoken;
+    const token = req.cookies.admintoken;
     if(!token) {
+        entryAccess = false
         return res.send('Failed to logout')
     }
     const data = jwt.verify(token, JWT_KEY_SECRET)
-    console.log("Logging out now")
-    return res.clearCookie('access_token')
-        .redirect('/admin/login');
+    console.log("Logging out now", data)
+    return res.clearCookie('admintoken').redirect('/admin/login');
 }
 
 
 
 // ADMIN HOME PAGE
 const adminHome = async (req, res) => {
-    let accessGranted = false
-    if(req.cookies.access_token) {
-        accessGranted = true
-    } 
-    const kloks = await Klok.find({});
-    const users = await User.find({});
-    const projects = await Project.find({});
-    res.render("admin/admin.ejs", { users, projects, kloks, accessGranted})
+        let entryAccess = false;
+        if (req.cookies.admintoken) {
+            entryAccess = true
+            const kloks = await Klok.find({});
+            const users = await User.find({});
+            const projects = await Project.find({});
+            res.render("admin/admin.ejs", { users, projects, kloks, entryAccess})
+        } else {
+            res.send("error: no access granted")
+        }
 }
 
 
 
 // CREATE NEW USER
-
 const createUser = async (req, res) => {
-    const requiredFields = ['username', 'fullname', 'PIN', 'startDate', 'position', 'hourlyPay'];
-    for(let i=0; i < requiredFields.length; i++) {
-        const field = requiredFields[i]
+    let entryAccess = req.cookies.admintoken
+    if (entryAccess) {
+        const requiredFields = ['username', 'fullname', 'PIN', 'startDate', 'position', 'hourlyPay'];
+        for(let field of requiredFields) {
         if(!(field in req.body)) {
             const errorMessage = `missing ${field} in request body`;
             return res.send(errorMessage);
-    }}
-    req.body.username = req.body.username.toLowerCase()
-    const { username, fullname, PIN, startDate, position, hourlyPay } = req.body
-    bcrypt.hash(PIN, 12)
-        .then(encryptedPIN => {
-            const userParams = { 
-                username, fullname,
-                PIN: encryptedPIN, 
-                startDate, position, hourlyPay
-            }
-            User.create(userParams)
-                .then(user => {
-                    const token = jwt.sign(
-                        { userId: user.id, PIN: user.PIN },
-                        JWT_KEY_SECRET, 
-                    )
-                    return res.cookie('access_token', token).redirect('/admin/home')
-                        
-            })
-        }
-        )
+        }}
+        const { username, fullname, PIN, startDate, position, hourlyPay } = req.body
+        try {
+            const newUser = { username, fullname, PIN, startDate, position, hourlyPay }
+            const user = User.create(newUser)
+            const token = jwt.sign(
+                { userId: user.id, PIN: user.PIN },
+                JWT_KEY_SECRET,
+                { expiresIn: '1hr' }
+                );
+            return res.cookie('admintoken', token).redirect('/admin/home')  
+        } catch (error) {
+            console.error(error)
+        } 
+} else {
+    res.send("error: no access granted")
+}
 }
 
 
+
 const createKlokAdmin = async (req, res) => {
-    const newKlok = await Klok.create(req.body)
-    res.redirect("/admin/home", { newKlok })
+    let entryAccess = req.cookies.admintoken
+    if (entryAccess) {
+        const newKlok = await Klok.create(req.body)
+        res.redirect("/admin/home", { newKlok })
+    } else {
+        res.send("error: no access granted")
+    }
 }
 
 
 const createProject = async (req, res) => {
+    let entryAccess = !req.cookies.admintoken
+    if (!entryAccess) {
     const newProj = await Project.create(req.body)
     res.redirect("/admin/home/", { newProj })
+    } else {
+        res.send("error: no access granted")
+    }
 }
 
 
@@ -194,11 +197,19 @@ const findAllProjects = async (req, res) => {
 // // UPDATE
 
 const  updateUserForm = async (req, res) => {
-    const user = await User.findOne({_id: req.params.id});
-    res.render("admin/edituser", { user })
+    let entryAccess = !req.cookies.admintoken
+    if (!entryAccess) {
+        const user = await User.findOne({_id: req.params.id});
+        res.render("admin/edituser", { user, entryAccess })
+    } else {
+        res.send("error: no access granted")
+    }
+    
 }
 
 const updateUser = async (req, res) => {
+    let entryAccess = !req.cookies.admintoken
+    if (!entryAccess) {
     const user = await User.findOneAndUpdate(
         {_id: req.params.id},
         {$set: req.body},
@@ -206,14 +217,24 @@ const updateUser = async (req, res) => {
         )
         console.log(user);
         res.redirect("/admin/home")
+    } else {
+        res.send("error: no access granted")
+    }
 }
 
 const  updateProjectForm = async (req, res) => {
-    const project = await Project.findOne({_id: req.params._id});
-    res.render("admin/editproject", { project })
+    let entryAccess = !req.cookies.admintoken
+    if (!entryAccess) {
+        const project = await Project.findOne({_id: req.params.id});
+        res.render("admin/editproject", { project, entryAccess })
+    }  else {
+        res.send("error: no access granted")
+    }
 }
 
 const updateProject = async (req, res) => {
+    let entryAccess = !req.cookies.admintoken
+    if (!entryAccess) {
     const user = await Project.findOneAndUpdate(
         {_id: req.params.id},
         {$set: req.body},
@@ -221,14 +242,24 @@ const updateProject = async (req, res) => {
         )
         console.log(user);
         res.redirect("/admin/update-user/:id")
+    } else {
+        res.send("error: no access granted")
+    }
 }
 
 const  updateKlokForm = async (req, res) => {
-    const klok = await Klok.findOne({ _id:req.params._id });
-    res.render("admin/editklok", { klok })
+    let entryAccess = !req.cookies.admintoken
+    if (!entryAccess) {
+        const klok = await Klok.findOne({ _id:req.params._id });
+        res.render("admin/editklok", { klok })
+    } else {
+        res.send("error: no access granted")
+    }
 }
 
 const updateKlok = async (req, res) => {
+    let entryAccess = !req.cookies.admintoken
+    if (!entryAccess) {
     const klok = await Klok.findOneAndUpdate(
         {_id: req.params.id},
         {$set: req.body},
@@ -236,50 +267,44 @@ const updateKlok = async (req, res) => {
         )
         console.log(klok);
         res.redirect("/admin/home")
+    } else {
+        res.send("error: no access granted")
+    }
 }
 
 // // DELETE
 
 const deleteUser = async (req, res) => {
-    try {
+    let entryAccess = !req.cookies.admintoken
+    if (!entryAccess) {
         const deletedUser = await User.findByIdAndDelete(req.params.id);
-        if (deleteUser) {
-            console.log("Deleted Project:", deleteUser);
-            res.redirect("/admin/home");
-        } else {
-            res.status(404).send("User not found.");
-        }
-    } catch (error) {
-        res.status(500).send("Internal Server Error");
+        console.log("Deleted User:", deleteUser);
+        res.redirect("/admin/home");
+    } else {
+        res.send("error: no access granted")
     }
-}
+}    
 
 const deleteProject = async (req, res) => {
-    try {
+    let entryAccess = !req.cookies.admintoken
+    if (!entryAccess) {
         const deletedProj = await Project.findByIdAndDelete(req.params.id);
-        if (deleteProject) {
-            console.log("Deleted Project:", deletedProj);
-            res.redirect("/admin/home");
-        } else {
-            res.status(404).send("Project not found.");
-        }
-    } catch (error) {
-        res.status(500).send("Internal Server Error");
-    }
+        console.log("Deleted Project:", deletedProj);
+        res.redirect("/admin/home");
+    } else {
+        res.send("error: no access granted")
+}
 }
 
 const deleteKlok = async (req, res) => {
-    try {
+    let entryAccess = !req.cookies.admintoken
+    if (!entryAccess) {
         const deletedKlok = await Klok.findByIdAndDelete(req.params.id);
-        if (deleteKlok) {
-            console.log("Deleted Klok:", deletedKlok);
-            res.redirect("/admin/home");
-        } else {
-            res.status(404).send("Klok not found.");
-        }
-    } catch (error) {
-        res.status(500).send("Internal Server Error");
-    }
+        console.log("Deleted Klok:", deletedKlok);
+        res.redirect("/admin/home");
+    } else {
+        res.send("error: no access granted")
+}
 }
 
 module.exports = {
@@ -295,7 +320,6 @@ module.exports = {
     deleteProject,
     adminLogin,
     adminLogout,
-    homePage,
     sendLoginForm,
     sendNewAdminForm,
     createNewAdmin,
